refactor: :label: move tiptap-extensions to tiptap folder
